Move RegionTarget and RegionDeps where they into clean/auto_trait
